chore(remap): test cue doc examples
