+++ lib/vrl/tests/Cargo.toml
@@ -15,6 +15,7 @@ glob = "0.3"
 prettydiff = "0.4"
 regex = "1"
 serde_json = "1"
+serde = "1"
 structopt = { version = "0.3", default-features = false }
 
 [dependencies.stdlib]
