+++ lib/vrl/tests/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod docs;
+mod test;
+
+pub use test::Test;
